% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCI_ALMANAC_raw}
\alias{NCI_ALMANAC_raw}
\title{A \code{data.table} containing a subset of interesting drug combinations from the
NCI-ALMANAC}
\format{
\code{data.table} object
}
\usage{
data(NCI_ALMANAC_raw)
}
\description{
This is a subset of the NCI-ALMANAC drug combination dataset from
Holkeck et al. 2017 where the top 10 most synergistic and antagonistic
drug combinations were selected based on ranking via ranking with Bliss
score.
}
\references{
Holbeck, S. L., Camalier, R., Crowell, J. A., Govindharajulu, J. P., Hollingshead, M., Anderson, L. W., Polley, E., Rubinstein, L., Srivastava, A., Wilsker, D., Collins, J. M., & Doroshow, J. H. (2017). The National Cancer Institute ALMANAC: A Comprehensive Screening Resource for the Detection of Anticancer Drug Pairs with Enhanced Therapeutic Activity. Cancer Research, 77(13), 3564–3576. https://doi.org/10.1158/0008-5472.CAN-17-0489
}
\keyword{datasets}
